+++ src/hooks/useToken.js
@@ -1,5 +1,19 @@
 import {useState} from "react"
 
 const useToken = () => {
-    const [token, setToken] = useState()
+    const getToken= () => {
+        const tokenString = sessionStorage.getItem("token")
+        const userToken = JSON.parse(tokenString)
+        return userToken?.token
+      }
+    const [token, setToken] = useState(getToken())
+
+    const saveToken = userToken => {
+        sessionStorage.setItem("token", JSON.stringify(userToken))
+      }
+    return {
+        setToken: saveToken,
+        token
+    }
 }
+export default useToken
